pyproject.toml
[project]
authors = [{name = "mholeys"}]
dependencies = ["splat64[mips]>=0.39.0,<0.40"]
name = "hg2-splat"
requires-python = ">= 3.11"
version = "0.1.0"

[build-system]
build-backend = "hatchling.build"
requires = ["hatchling"]

[tool.pixi.workspace]
channels = ["conda-forge"]
platforms = ["linux-64"]

[tool.pixi.pypi-dependencies]
hg2_splat = { path = ".", editable = true }

[tool.pixi.tasks]
